feat: Complete card modularization and implement improved ESLint configuration
- Implement eslint-plugin-unused-imports for automatic unused import removal
- Modularize all card definitions into individual files under dreitron_exodia deck
- Create comprehensive extra monster cards including Aurora Dragon, Union Carrier, Link Riboh, etc.
- Delete centralized card definition files in favor of modular structure
- Add deckUtils for reusable deck building functions
- Update gameStore to use new dreitron_exodia deck structure
- Improve code organization with cleaner import management

Major changes:
- Split 9 common monsters, 8 magic cards, 1 trap card, 8 extra monsters, 1 token into individual files
- Each card now exports as default from dedicated file for better maintainability
- Enhanced ESLint configuration with unused import detection and auto-fix
- Removed obsolete centralized card definition system
